+++ include/omni/isaac/mjcf/mjcf.h
@@ -1,4 +1,11 @@
-// Copyright (c) 2022, NVIDIA CORPORATION. All rights reserved.
+// Copyright (c) 2022-2023, NVIDIA CORPORATION. All rights reserved.
+//
+// NVIDIA CORPORATION and its licensors retain all intellectual property
+// and proprietary rights in and to this software, related documentation
+// and any modifications thereto. Any use, reproduction, disclosure or
+// distribution of this software and related documentation without an express
+// license agreement from NVIDIA CORPORATION is strictly prohibited.
+//
 //
 // NVIDIA CORPORATION and its licensors retain all intellectual property
 // and proprietary rights in and to this software, related documentation
@@ -39,9 +46,12 @@ struct ImportConfig
     bool createBodyForFixedJoint = true;
     bool overrideCoM = false;
     bool overrideInertia = false;
+    bool visualizeCollisionGeoms = false;
+    bool importSites = true;
 
     bool makeInstanceable = false;
     std::string instanceableMeshUsdPath = "./instanceable_meshes.usd";
+    std::string meshRootDirectory = "";
 };
 
 struct Mjcf